requirements.txt
Flask
gunicorn
pyodbc
python-dotenv
scikit-learn
joblib
numpy
pandas